Add files via upload
[These files were for the Defense course]

Task 6 is the final report: an extensive SBAR regarding the fictional investigation of a suspicious (phishing) email sent to a CFO; the excel file consists of the timeline and list of IOCs. 

Tools used: Wireshark, Splunk, Autopsy, Volatility 

________________________________

Tasks 2-5 consisted of the preliminary work which resulted in Task 6. Task 1 was essentially a rough draft of an SBAR for an unrelated remote intrusion attempt.